[REFACTOR] DRY variety and batch selectors
